Use ProjectMetadataContext in ProjectHeader component
