+++ src/components/v2v3/V2V3Project/V2V3ProjectSettings/V2V3ProjectSettings.tsx
@@ -1,7 +1,7 @@
 import { LeftOutlined } from '@ant-design/icons'
 import { t, Trans } from '@lingui/macro'
 import { Button, Layout, Menu, MenuProps, Space } from 'antd'
-import ProjectHeader from 'components/Project/ProjectHeader'
+import { ProjectHeader } from 'components/Project/ProjectHeader'
 import { V2V3ProjectHeaderActions } from 'components/v2v3/V2V3Project/V2V3ProjectHeaderActions'
 import { ProjectSettingsContent } from 'components/v2v3/V2V3Project/V2V3ProjectSettings/ProjectSettingsContent'
 import { FEATURE_FLAGS } from 'constants/featureFlags'
@@ -85,9 +85,7 @@ const items: MenuItem[] = [
 export function V2V3ProjectSettings() {
   const { isPreviewMode, projectOwnerAddress, handle } =
     useContext(V2V3ProjectContext)
-  const { projectMetadata, isArchived, projectId } = useContext(
-    ProjectMetadataContext,
-  )
+  const { projectId } = useContext(ProjectMetadataContext)
   const { isDarkMode } = useContext(ThemeContext)
 
   const router = useRouter()
@@ -107,13 +105,10 @@ export function V2V3ProjectSettings() {
     <div style={layouts.maxWidth}>
       <Space direction="vertical" size={40} style={{ width: '100%' }}>
         <ProjectHeader
-          metadata={projectMetadata}
           actions={!isPreviewMode ? <V2V3ProjectHeaderActions /> : undefined}
-          isArchived={isArchived}
           handle={handle}
           projectOwnerAddress={projectOwnerAddress}
           canEditProjectHandle={canEditProjectHandle}
-          projectId={projectId}
         />
         <Layout
           style={{
